+++ packages/back-end/types/openapi.d.ts
@@ -146,10 +146,28 @@ export interface paths {
   "/visual-changesets/{id}": {
     /** Get a single visual changeset */
     get: operations["getVisualChangeset"];
+    /** Update a visual changeset */
+    put: operations["putVisualChangeset"];
+  };
+  "/visual-changesets/{id}/visual-change": {
+    /** Create a visual change for a visual changeset */
+    post: operations["postVisualChange"];
     parameters: {
         /** @description The id of the requested resource */
       path: {
         id: string;
+      };
+    };
+  };
+  "/visual-changesets/{id}/visual-change/{visualChangeId}": {
+    /** Update a visual change for a visual changeset */
+    put: operations["putVisualChange"];
+    parameters: {
+        /** @description The id of the requested resource */
+        /** @description Specify a specific visual change */
+      path: {
+        id: string;
+        visualChangeId: string;
       };
     };
   };
@@ -544,6 +562,8 @@ export interface components {
       project: string;
       encryptPayload: boolean;
       encryptionKey: string;
+      includeVisualExperiments?: boolean;
+      includeDraftExperiments?: boolean;
       key: string;
       proxyEnabled: boolean;
       proxyHost: string;
@@ -802,7 +822,12 @@ export interface components {
     };
     VisualChangeset: {
       id?: string;
-      urlPattern: string;
+      urlPatterns: ({
+          include?: boolean;
+          /** @enum {string} */
+          type: "simple" | "regex";
+          pattern: string;
+        })[];
       editorUrl: string;
       experiment: string;
       visualChanges: ({
@@ -816,6 +841,18 @@ export interface components {
               attribute: string;
               value?: string;
             })[];
+        })[];
+    };
+    VisualChange: {
+      description?: string;
+      css?: string;
+      variation: string;
+      domMutations?: ({
+          selector: string;
+          /** @enum {string} */
+          action: "append" | "set" | "remove";
+          attribute: string;
+          value?: string;
         })[];
     };
   };
@@ -833,6 +870,8 @@ export interface components {
     projectId: string;
     /** @description Filter by Data Source */
     datasourceId: string;
+    /** @description Specify a specific visual change */
+    visualChangeId: string;
   };
   requestBodies: never;
   headers: never;
@@ -1427,6 +1466,8 @@ export interface operations {
                 project: string;
                 encryptPayload: boolean;
                 encryptionKey: string;
+                includeVisualExperiments?: boolean;
+                includeDraftExperiments?: boolean;
                 key: string;
                 proxyEnabled: boolean;
                 proxyHost: string;
@@ -1462,6 +1503,8 @@ export interface operations {
               project: string;
               encryptPayload: boolean;
               encryptionKey: string;
+              includeVisualExperiments?: boolean;
+              includeDraftExperiments?: boolean;
               key: string;
               proxyEnabled: boolean;
               proxyHost: string;
@@ -2213,7 +2256,12 @@ export interface operations {
           "application/json": {
             visualChangesets: ({
                 id?: string;
-                urlPattern: string;
+                urlPatterns: ({
+                    include?: boolean;
+                    /** @enum {string} */
+                    type: "simple" | "regex";
+                    pattern: string;
+                  })[];
                 editorUrl: string;
                 experiment: string;
                 visualChanges: ({
@@ -2236,13 +2284,28 @@ export interface operations {
   };
   getVisualChangeset: {
     /** Get a single visual changeset */
+    parameters: {
+        /** @description Include the associated experiment in payload */
+      query: {
+        includeExperiment?: number;
+      };
+        /** @description The id of the requested resource */
+      path: {
+        id: string;
+      };
+    };
     responses: {
       200: {
         content: {
           "application/json": {
             visualChangeset: {
               id?: string;
-              urlPattern: string;
+              urlPatterns: ({
+                  include?: boolean;
+                  /** @enum {string} */
+                  type: "simple" | "regex";
+                  pattern: string;
+                })[];
               editorUrl: string;
               experiment: string;
               visualChanges: ({
@@ -2258,6 +2321,135 @@ export interface operations {
                     })[];
                 })[];
             };
+            experiment?: {
+              id: string;
+              /** Format: date-time */
+              dateCreated: string;
+              /** Format: date-time */
+              dateUpdated: string;
+              name: string;
+              project: string;
+              hypothesis: string;
+              description: string;
+              tags: (string)[];
+              owner: string;
+              archived: boolean;
+              status: string;
+              autoRefresh: boolean;
+              hashAttribute: string;
+              variations: ({
+                  variationId: string;
+                  key: string;
+                  name: string;
+                  description: string;
+                  screenshots: (string)[];
+                })[];
+              phases: ({
+                  name: string;
+                  dateStarted: string;
+                  dateEnded: string;
+                  reasonForStopping: string;
+                  seed: string;
+                  coverage: number;
+                  trafficSplit: ({
+                      variationId: string;
+                      weight: number;
+                    })[];
+                  namespace?: {
+                    namespaceId: string;
+                    range: (unknown)[];
+                  };
+                  targetingCondition: string;
+                })[];
+              settings: {
+                datasourceId: string;
+                assignmentQueryId: string;
+                experimentId: string;
+                segmentId: string;
+                queryFilter: string;
+                /** @enum {unknown} */
+                inProgressConversions: "include" | "exclude";
+                /** @enum {unknown} */
+                attributionModel: "firstExposure" | "experimentDuration";
+                /** @enum {unknown} */
+                statsEngine: "bayesian" | "frequentist";
+                goals: ({
+                    metricId: string;
+                    overrides: {
+                      conversionWindowStart?: number;
+                      conversionWindowEnd?: number;
+                      winRiskThreshold?: number;
+                      loseRiskThreshold?: number;
+                    };
+                  })[];
+                guardrails: ({
+                    metricId: string;
+                    overrides: {
+                      conversionWindowStart?: number;
+                      conversionWindowEnd?: number;
+                      winRiskThreshold?: number;
+                      loseRiskThreshold?: number;
+                    };
+                  })[];
+                activationMetric?: {
+                  metricId: string;
+                  overrides: {
+                    conversionWindowStart?: number;
+                    conversionWindowEnd?: number;
+                    winRiskThreshold?: number;
+                    loseRiskThreshold?: number;
+                  };
+                };
+              };
+              resultSummary?: {
+                status: string;
+                winner: string;
+                conclusions: string;
+                releasedVariationId: string;
+              };
+            };
+          };
+        };
+      };
+    };
+  };
+  putVisualChangeset: {
+    /** Update a visual changeset */
+    parameters: {
+        /** @description The id of the requested resource */
+      path: {
+        id: string;
+      };
+    };
+    responses: {
+      200: {
+        content: {
+          "application/json": {
+            nModified: number;
+          };
+        };
+      };
+    };
+  };
+  postVisualChange: {
+    /** Create a visual change for a visual changeset */
+    responses: {
+      200: {
+        content: {
+          "application/json": {
+            nModified: number;
+          };
+        };
+      };
+    };
+  };
+  putVisualChange: {
+    /** Update a visual change for a visual changeset */
+    responses: {
+      200: {
+        content: {
+          "application/json": {
+            nModified: number;
           };
         };
       };
@@ -2285,6 +2477,7 @@ export type ApiExperimentAnalysisSettings = components["schemas"]["ExperimentAna
 export type ApiExperimentResults = components["schemas"]["ExperimentResults"];
 export type ApiDataSource = components["schemas"]["DataSource"];
 export type ApiVisualChangeset = components["schemas"]["VisualChangeset"];
+export type ApiVisualChange = components["schemas"]["VisualChange"];
 
 // Operations
 export type ListFeaturesResponse = operations["listFeatures"]["responses"]["200"]["content"]["application/json"];
@@ -2308,3 +2501,6 @@ export type PostMetricResponse = operations["postMetric"]["responses"]["200"]["c
 export type GetMetricResponse = operations["getMetric"]["responses"]["200"]["content"]["application/json"];
 export type ListVisualChangesetsResponse = operations["listVisualChangesets"]["responses"]["200"]["content"]["application/json"];
 export type GetVisualChangesetResponse = operations["getVisualChangeset"]["responses"]["200"]["content"]["application/json"];
+export type PutVisualChangesetResponse = operations["putVisualChangeset"]["responses"]["200"]["content"]["application/json"];
+export type PostVisualChangeResponse = operations["postVisualChange"]["responses"]["200"]["content"]["application/json"];
+export type PutVisualChangeResponse = operations["putVisualChange"]["responses"]["200"]["content"]["application/json"];